#9758: Moving pool pybind
