+++ tt_eager/tt_lib/csrc/tt_lib_bindings_tensor.cpp
@@ -18,8 +18,6 @@
 #include "tt_dnn/op_library/fully_connected/fully_connected_op.hpp"
 #include "tt_dnn/op_library/groupnorm/groupnorm_op.hpp"
 #include "tt_dnn/op_library/layernorm/layernorm_op.hpp"
-#include "ttnn/cpp/ttnn/operations/pool/average_pool.hpp"
-#include "ttnn/cpp/ttnn/operations/pool/max_pool.hpp"
 #include "tt_dnn/op_library/reduce/reduce_op.hpp"
 #include "tt_dnn/op_library/fast_reduce_nc/fast_reduce_nc_op.hpp"
 #include "tt_dnn/op_library/rotary_embedding/rotary_embedding_op.hpp"
@@ -718,22 +716,22 @@ void TensorModule(py::module& m_tensor) {
     )doc");
 
     // Pools
-    m_tensor.def(
-        "average_pool_2d",
-        &average_pool_2d,
-        py::arg().noconvert(),
-        py::arg("output_mem_config").noconvert() = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-        py::arg("output_dtype").noconvert() = std::nullopt,
-        R"doc(
-        Average Pool 2D
-        It operates on tensors whose that have channels as the last dimension
+    // m_tensor.def(
+    //     "average_pool_2d",
+    //     &average_pool_2d,
+    //     py::arg().noconvert(),
+    //     py::arg("output_mem_config").noconvert() = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
+    //     py::arg("output_dtype").noconvert() = std::nullopt,
+    //     R"doc(
+    //     Average Pool 2D
+    //     It operates on tensors whose that have channels as the last dimension
 
-        +----------+----------------------------+------------+-------------------------------+----------+
-        | Argument | Description                | Data type  | Valid range                   | Required |
-        +==========+============================+============+===============================+==========+
-        | act      | Input activations tensor   | Tensor     |                               | Yes      |
-        +----------+----------------------------+------------+-------------------------------+----------+
-    )doc");
+    //     +----------+----------------------------+------------+-------------------------------+----------+
+    //     | Argument | Description                | Data type  | Valid range                   | Required |
+    //     +==========+============================+============+===============================+==========+
+    //     | act      | Input activations tensor   | Tensor     |                               | Yes      |
+    //     +----------+----------------------------+------------+-------------------------------+----------+
+    // )doc");
 
     // Upsample
     m_tensor.def(
@@ -754,84 +752,84 @@ void TensorModule(py::module& m_tensor) {
         +----------+----------------------------+------------+-------------------------------+----------+
     )doc");
 
-    m_tensor.def(
-        "max_pool2d",
-        &max_pool2d,
-        py::arg("input").noconvert(),
-        py::arg("in_n").noconvert(),
-        py::arg("in_h").noconvert(),
-        py::arg("in_w").noconvert(),
-        py::arg("kernel_h").noconvert(),
-        py::arg("kernel_w").noconvert(),
-        py::arg("stride_h") = 1,
-        py::arg("stride_w") = 1,
-        py::arg("pad_h") = 0,
-        py::arg("pad_w") = 0,
-        py::arg("dilation_h") = 1,
-        py::arg("dilation_w") = 1,
-        py::arg("output_mem_config") = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-        py::arg("nblocks") = 1,
-        py::arg("use_multicore") = true,
-        R"doc(
-        Max Pool 2D
-        +-------------------+-------------------------------+---------------+-------------+----------+
-        | Argument          | Description                   | Data type     | Valid range | Required |
-        +===================+===============================+===============+=============+==========+
-        | input             | Input activations tensor      | Tensor        |             | Yes      |
-        | in_n              | Input nbatch                  | Tensor        |             | Yes      |
-        | in_h              | Input height                  | Tensor        |             | Yes      |
-        | in_w              | Input width                   | Tensor        |             | Yes      |
-        | kernel_h          | kernel window height          | uint32_t      |             | Yes      |
-        | kernel_w          | kernel window width           | uint32_t      |             | Yes      |
-        | stride_h          | stride in height dim          | uint32_t      |             | No       |
-        | stride_w          | stride in width dim           | uint32_t      |             | No       |
-        | pad_h             | padding in height dim         | uint32_t      |             | No       |
-        | pad_w             | padding in width dim          | uint32_t      |             | No       |
-        | dilation_h        | kernel dilation in height dim | uint32_t      |             | No       |
-        | dilation_w        | kernel dilation in width dim  | uint32_t      |             | No       |
-        | output_mem_config | output tensor memory config   | MemoryConfig  |             | No       |
-        +-------------------+-------------------------------+---------------+-------------+----------+
-    )doc");
+    // m_tensor.def(
+    //     "max_pool2d",
+    //     &max_pool2d,
+    //     py::arg("input").noconvert(),
+    //     py::arg("in_n").noconvert(),
+    //     py::arg("in_h").noconvert(),
+    //     py::arg("in_w").noconvert(),
+    //     py::arg("kernel_h").noconvert(),
+    //     py::arg("kernel_w").noconvert(),
+    //     py::arg("stride_h") = 1,
+    //     py::arg("stride_w") = 1,
+    //     py::arg("pad_h") = 0,
+    //     py::arg("pad_w") = 0,
+    //     py::arg("dilation_h") = 1,
+    //     py::arg("dilation_w") = 1,
+    //     py::arg("output_mem_config") = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
+    //     py::arg("nblocks") = 1,
+    //     py::arg("use_multicore") = true,
+    //     R"doc(
+    //     Max Pool 2D
+    //     +-------------------+-------------------------------+---------------+-------------+----------+
+    //     | Argument          | Description                   | Data type     | Valid range | Required |
+    //     +===================+===============================+===============+=============+==========+
+    //     | input             | Input activations tensor      | Tensor        |             | Yes      |
+    //     | in_n              | Input nbatch                  | Tensor        |             | Yes      |
+    //     | in_h              | Input height                  | Tensor        |             | Yes      |
+    //     | in_w              | Input width                   | Tensor        |             | Yes      |
+    //     | kernel_h          | kernel window height          | uint32_t      |             | Yes      |
+    //     | kernel_w          | kernel window width           | uint32_t      |             | Yes      |
+    //     | stride_h          | stride in height dim          | uint32_t      |             | No       |
+    //     | stride_w          | stride in width dim           | uint32_t      |             | No       |
+    //     | pad_h             | padding in height dim         | uint32_t      |             | No       |
+    //     | pad_w             | padding in width dim          | uint32_t      |             | No       |
+    //     | dilation_h        | kernel dilation in height dim | uint32_t      |             | No       |
+    //     | dilation_w        | kernel dilation in width dim  | uint32_t      |             | No       |
+    //     | output_mem_config | output tensor memory config   | MemoryConfig  |             | No       |
+    //     +-------------------+-------------------------------+---------------+-------------+----------+
+    // )doc");
 
-    m_tensor.def(
-        "max_pool2d_v2",
-        &max_pool2d_v2,
-        py::arg("input").noconvert(),
-        py::arg("reader_indices").noconvert(),
-        py::arg("in_n").noconvert(),
-        py::arg("in_h").noconvert(),
-        py::arg("in_w").noconvert(),
-        py::arg("kernel_h").noconvert(),
-        py::arg("kernel_w").noconvert(),
-        py::arg("stride_h") = 1,
-        py::arg("stride_w") = 1,
-        py::arg("pad_h") = 0,
-        py::arg("pad_w") = 0,
-        py::arg("dilation_h") = 1,
-        py::arg("dilation_w") = 1,
-        py::arg("output_mem_config") = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-        py::arg("nblocks") = 1,
-        py::arg("use_multicore") = true,
-        R"doc(
-        Max Pool 2D
-        +-------------------+-------------------------------+---------------+-------------+----------+
-        | Argument          | Description                   | Data type     | Valid range | Required |
-        +===================+===============================+===============+=============+==========+
-        | input             | Input activations tensor      | Tensor        |             | Yes      |
-        | in_n              | Input nbatch                  | Tensor        |             | Yes      |
-        | in_h              | Input height                  | Tensor        |             | Yes      |
-        | in_w              | Input width                   | Tensor        |             | Yes      |
-        | kernel_h          | kernel window height          | uint32_t      |             | Yes      |
-        | kernel_w          | kernel window width           | uint32_t      |             | Yes      |
-        | stride_h          | stride in height dim          | uint32_t      |             | No       |
-        | stride_w          | stride in width dim           | uint32_t      |             | No       |
-        | pad_h             | padding in height dim         | uint32_t      |             | No       |
-        | pad_w             | padding in width dim          | uint32_t      |             | No       |
-        | dilation_h        | kernel dilation in height dim | uint32_t      |             | No       |
-        | dilation_w        | kernel dilation in width dim  | uint32_t      |             | No       |
-        | output_mem_config | output tensor memory config   | MemoryConfig  |             | No       |
-        +-------------------+-------------------------------+---------------+-------------+----------+
-    )doc");
+    // m_tensor.def(
+    //     "max_pool2d_v2",
+    //     &max_pool2d_v2,
+    //     py::arg("input").noconvert(),
+    //     py::arg("reader_indices").noconvert(),
+    //     py::arg("in_n").noconvert(),
+    //     py::arg("in_h").noconvert(),
+    //     py::arg("in_w").noconvert(),
+    //     py::arg("kernel_h").noconvert(),
+    //     py::arg("kernel_w").noconvert(),
+    //     py::arg("stride_h") = 1,
+    //     py::arg("stride_w") = 1,
+    //     py::arg("pad_h") = 0,
+    //     py::arg("pad_w") = 0,
+    //     py::arg("dilation_h") = 1,
+    //     py::arg("dilation_w") = 1,
+    //     py::arg("output_mem_config") = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
+    //     py::arg("nblocks") = 1,
+    //     py::arg("use_multicore") = true,
+    //     R"doc(
+    //     Max Pool 2D
+    //     +-------------------+-------------------------------+---------------+-------------+----------+
+    //     | Argument          | Description                   | Data type     | Valid range | Required |
+    //     +===================+===============================+===============+=============+==========+
+    //     | input             | Input activations tensor      | Tensor        |             | Yes      |
+    //     | in_n              | Input nbatch                  | Tensor        |             | Yes      |
+    //     | in_h              | Input height                  | Tensor        |             | Yes      |
+    //     | in_w              | Input width                   | Tensor        |             | Yes      |
+    //     | kernel_h          | kernel window height          | uint32_t      |             | Yes      |
+    //     | kernel_w          | kernel window width           | uint32_t      |             | Yes      |
+    //     | stride_h          | stride in height dim          | uint32_t      |             | No       |
+    //     | stride_w          | stride in width dim           | uint32_t      |             | No       |
+    //     | pad_h             | padding in height dim         | uint32_t      |             | No       |
+    //     | pad_w             | padding in width dim          | uint32_t      |             | No       |
+    //     | dilation_h        | kernel dilation in height dim | uint32_t      |             | No       |
+    //     | dilation_w        | kernel dilation in width dim  | uint32_t      |             | No       |
+    //     | output_mem_config | output tensor memory config   | MemoryConfig  |             | No       |
+    //     +-------------------+-------------------------------+---------------+-------------+----------+
+    // )doc");
 
     // TMs
     m_tensor.def(
