+++ ttnn/cpp/pybind11/operations/__init__.hpp
@@ -16,12 +16,12 @@
 #include "pybind11/operations/kv_cache.hpp"
 #include "pybind11/operations/matmul.hpp"
 #include "pybind11/operations/normalization.hpp"
-#include "pybind11/operations/pool.hpp"
 #include "pybind11/operations/copy.hpp"
 #include "pybind11/operations/ternary.hpp"
 #include "pybind11/operations/transformer.hpp"
 #include "pybind11/operations/unary.hpp"
 
+#include "ttnn/operations/pool/pool_pybind.hpp"
 #include "ttnn/operations/eltwise/binary/binary_pybind.hpp"
 #include "ttnn/operations/reduction/reduction_pybind.hpp"
 #include "ttnn/operations/eltwise/binary_backward/binary_backward_pybind.hpp"
